+++ src/pages/SearchPage.js
@@ -3,6 +3,7 @@ import { Outlet } from "react-router-dom";
 import FilterTab from "../components/filterTab/FilterTab";
 import MoviesFound from "../components/moviesList/MoviesFound";
 import MoviesList from "../components/moviesList/MoviesList";
+import{ movies} from "../constants/movies";
 
 const SearchPage = () => (
   <div className="flex flex-col h-screen bg-gray-200">
@@ -10,7 +11,7 @@ const SearchPage = () => (
     <div className="bg-gray-500 mt-2 h-max px-10 py-1">
       <FilterTab />
       <MoviesFound />
-      <MoviesList />
+      <MoviesList movies={movies} />
     </div>
     <Footer />
   </div>
